Migrate long consumptions to double

+++ SPBackend/SPBackend/DTOs/RoomConsumption.cs
@@ -5,5 +5,5 @@ public class RoomConsumption
     public long RoomId { get; set; }
     public string Name { get; set; }
     public string RoomType { get; set; }
-    public long Consumption { get; set; }
+    public double Consumption { get; set; }
 }
