+++ SPBackend/SPBackend/Models/RecentConsumption.cs
@@ -4,7 +4,7 @@ public class RecentConsumption
 {
     public long Id { get; set; }
     public DateTime Time { get; set; }
-    public long TotalEnergy { get; set; }
+    public double TotalEnergy { get; set; }
     public long PlugId { get; set; }
     public Plug Plug { get; set; }
 }
